Updated everything related to records
Now records work properly. Creating, updating, deleting etc.

There's a link to every record inside the brotherhood display to satisfy a requirement.

+++ src/main/java/controllers/brotherhood/EnrolmentBrotherhoodController.java
@@ -146,6 +146,10 @@ public class EnrolmentBrotherhoodController extends AbstractController {
 		enrolments = b.getEnrolments();
 
 		enrolment = this.enrolmentService.findOne(varId);
+
+		if (enrolment.getBrotherhood().getId() != this.actorService.findByPrincipal().getId())
+			return new ModelAndView("redirect:/welcome/index.do");
+
 		try {
 			this.enrolmentService.delete(enrolment);
 			enrolments = b.getEnrolments();
@@ -173,6 +177,10 @@ public class EnrolmentBrotherhoodController extends AbstractController {
 		enrolments = b.getEnrolments();
 
 		enrolment = this.enrolmentService.findOne(varId);
+
+		if (enrolment.getBrotherhood().getId() != this.actorService.findByPrincipal().getId())
+			return new ModelAndView("redirect:/welcome/index.do");
+
 		try {
 			this.enrolmentService.drop(enrolment);
 			enrolments = b.getEnrolments();
